requirements.txt
<<<<<<< HEAD
Django>=5.1,<5.2
wagtail>=6.3,<6.4
anyascii==0.3.2
asgiref==3.8.1
bcp47==0.0.4
beautifulsoup4==4.11.2
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
defusedxml==0.7.1
django-browser-reload==1.17.0
django-crispy-forms==2.3
django-filter==23.5
django-modelcluster==6.3
django-permissionedforms==0.1
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
django-taggit==6.1.0
django-treebeard==4.7.1
django-vite==3.0.5
djangorestframework==3.15.2
draftjs-exporter==2.1.7
et-xmlfile==2.0.0
ffmpeg==1.4
filetype==1.2.0
flake8==7.1.1
html5lib==1.1
=======
anyascii==0.3.2
asgiref==3.8.1
babel==2.17.0
bcp47==0.0.4
beautifulsoup4==4.11.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
crispy-bootstrap5==2024.10
cryptography==44.0.0
decorator==5.1.1
defusedxml==0.7.1
django==4.2.19
django-bootstrap5==24.3
django-browser-reload==1.17.0
django-crispy-forms==2.3
django-environ==0.11.2
django-eventtools==1.0.3
django-extra-views==0.14.0
django-filter==23.5
django-haystack==3.3.0
django-modelcluster==6.3
django-oscar==3.2.5
django-oscar-stripe-sca==0.8
django-permissionedforms==0.1
django-phonenumber-field==6.4.0
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
django-tables2==2.3.4
django-taggit==6.1.0
django-treebeard==4.7.1
django-vite==3.0.5
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
draftjs-exporter==2.1.7
et-xmlfile==2.0.0
factory-boy==3.2.1
faker==35.2.0
ffmpeg==1.4
filetype==1.2.0
flake8==7.1.1
future==1.0.0
geocoder==1.38.1
git-filter-repo==2.47.0
html5lib==1.1
icalendar==6.0.1
>>>>>>> 3a2d8aa8c1b0f4975ce54c17064465ab66e104bd
idna==3.10
l18n==2021.3
laces==0.1.1
mccabe==0.7.0
<<<<<<< HEAD
openpyxl==3.1.5
pillow==10.4.0
pillow-heif==0.20.0
pycodestyle==2.12.1
pyflakes==3.2.0
pymysql==1.1.1
pytz==2024.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.2
telepath==0.3.1
unidecode==1.3.8
urllib3==2.2.3
wagtail-django-recaptcha==2.1.1
wagtail-generic-chooser==0.6.1
wagtail-modeladmin==2.1.0
wagtailvideos==7.0.2
webencodings==0.5.1
willow==1.9.0
django-environ==0.11.2
=======
mysqlclient==2.2.6
openpyxl==3.1.5
packaging==24.2
phonenumbers==8.13.54
pillow==10.4.0
pillow-heif==0.20.0
purl==1.6
pycodestyle==2.12.1
pycountry==24.6.1
pycparser==2.22
pyflakes==3.2.0
pymysql==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
ratelim==0.1.6
requests==2.32.3
ruff==0.9.4
setuptools==75.8.0
six==1.17.0
sorl-thumbnail==12.9.0
soupsieve==2.6
sqlparse==0.5.2
stripe==11.5.0
telepath==0.3.1
typing-extensions==4.12.2
tzdata==2024.2
unidecode==1.3.8
urllib3==2.2.3
wagtail==6.3.1
wagtail-cache==2.5.1
wagtail-django-recaptcha==2.1.1
wagtail-generic-chooser==0.6.1
wagtail-modeladmin==2.1.0
wagtail-seo==2.5.0
wagtailvideos==7.0.2
webencodings==0.5.1
whoosh==2.7.4
willow==1.9.0
>>>>>>> 3a2d8aa8c1b0f4975ce54c17064465ab66e104bd
